Refactor auth-modal
