+++ packages/designmanual/stories/molecules/mastheads.jsx
@@ -23,7 +23,7 @@ import {
   TopicMenuButton,
   SearchFieldForm,
   BreadcrumbBlock,
-  AuthModal,
+  MastheadAuthModal,
 } from '@ndla/ui';
 import Modal from '@ndla/modal';
 import SafeLink from '@ndla/safelink';
@@ -227,7 +227,7 @@ class MastheadWithTopicMenu extends Component {
           <DisplayOnPageYOffset yOffsetMin={0} yOffsetMax={150}>
             <LanguageSelector inverted={ndlaFilm} options={dummyLanguageOptions} currentLanguage={i18n.language} />
           </DisplayOnPageYOffset>
-          <AuthModal inverted={ndlaFilm} authorizedRole="lærer" authorizedCollectedInfo={['Lærer', 'Skole']} />
+          <MastheadAuthModal inverted={ndlaFilm} authorizedRole="lærer" authorizedCollectedInfo={['Lærer', 'Skole']} />
           {this.renderSearchButtonView(true, ndlaFilm)}
           <Logo
             to="?selectedKind=Emnesider&selectedStory=1.%20Fagoversikt&full=0&addons=0&stories=1&panelRight=0&addonPanel=storybook%2Factions%2Factions-panel"
